(DEAD_END) demo logtag usage
THIS IS A DEAD-END BRANCH. DO NOT MAKE COMMITS INTO THIS BRANCH.

LogTag usage is added to three source files in opencv_core:
ocl.cpp, datafile.cpp, filesystem.cpp.

These are intended to demonstrate the use of logtags while minimizing
code changes. In particular, we simply cause existing CV_LOG_*(NULL)
to use CV_LOGTAG_FALLBACK within a compilation unit (cpp file).

These changes are not intended for merging. This commit will be
excluded on the next rebase.

THIS IS A DEAD-END BRANCH. DO NOT MAKE COMMITS INTO THIS BRANCH.